Patches
I'm tired boss.
Changes:
- Demidragon leader clothes layering fixed (again)
- Added names to new Lamia sliders
- Added Anthro Lamia sprites
- Adjusted Anthro Lamia layering
- Fixed Unit Tag loader function
- Added a way to view a unit's currently equipped potions
- Added a way to view a unit's currently active tags.

+++ Assets/Scripts/UI/Connectors/ConditionalTraitArithmeticPanel.cs
@@ -245,7 +245,7 @@ public class ConditionalTraitArithmeticPanel : MonoBehaviour
         if (current_mode != VariableMode.None)
         {
             currentPrefab.associetedBlock.compareOp = (TraitConditionCompareOperator)conditionOp.value;
-            currentPrefab.associetedBlock.compareValue = int.TryParse(conditionValue.text, out int cv) ? cv : -1;
+            currentPrefab.associetedBlock.compareValue = float.TryParse(conditionValue.text, out float cv) ? cv : -1;
             currentPrefab.associetedBlock.conditionVariable.Clear();
             foreach (var item in conditionVariable)
             {
